Move Lock to cdi package
